+++ extensions/projects/src/projects.tsx
@@ -1,17 +1,12 @@
 import { getPreferenceValues, Icon, List, showToast, Toast } from "@raycast/api";
 import { useState, ReactElement } from "react";
 import { SearchProjectActionPanel } from "./action-panel";
-import { Preferences } from "./types";
+import { ListType, Preferences, SourceRepo } from "./types";
 import { tildifyPath, useRepoCache } from "./projects-service";
+import applicationConfig from "./application-config.json";
 
 export default function Main(): ReactElement {
-  const searchBarPlaceholders = [
-    "Search for a project",
-    "Search projects by name",
-    "Search projects by type. e.g. 'type: node'",
-    "Search projects in directory. e.g. 'dir: ajay'",
-    "Combine keywords to search. e.g. 'dir: ajay type: node'",
-  ];
+  const searchPlaceholders = applicationConfig.searchPlaceholders;
 
   const preferences = getPreferenceValues<Preferences>();
   const [searchText, setSearchText] = useState<string>();
@@ -21,54 +16,34 @@ export default function Main(): ReactElement {
     showToast(Toast.Style.Failure, "", error);
   }
 
+  const getListSection = (listType: ListType, section: { sectionTitle: string; repos: SourceRepo[] }) => {
+    return (
+      <List.Section title={section?.sectionTitle}>
+        {section?.repos?.map((repo) => (
+          <List.Item
+            key={repo.fullPath}
+            id={`${listType}:${repo.id}`}
+            title={repo.name}
+            icon={repo.icon}
+            accessoryTitle={tildifyPath(repo.fullPath)}
+            keywords={[repo.name]}
+            accessoryIcon={listType == "pinned" ? Icon.Pin : ""}
+            actions={<SearchProjectActionPanel repo={repo} preferences={preferences} listType={listType} />}
+          />
+        ))}
+      </List.Section>
+    );
+  };
+
   return (
     <List
-      searchBarPlaceholder={searchBarPlaceholders[Math.floor(Math.random() * searchBarPlaceholders.length)]}
+      searchBarPlaceholder={searchPlaceholders[Math.floor(Math.random() * searchPlaceholders.length)]}
       onSearchTextChange={setSearchText}
       isLoading={isLoading}
     >
-      <List.Section title={response?.pinned?.sectionTitle}>
-        {response?.pinned?.repos?.map((repo) => (
-          <List.Item
-            key={repo.fullPath}
-            id={repo.id}
-            title={repo.name}
-            icon={repo.icon}
-            accessoryTitle={tildifyPath(repo.fullPath)}
-            keywords={[repo.name]}
-            accessoryIcon={Icon.Pin}
-            actions={<SearchProjectActionPanel repo={repo} preferences={preferences} pinned />}
-          />
-        ))}
-      </List.Section>
-
-      <List.Section title={response?.recent?.sectionTitle}>
-        {response?.recent?.repos?.map((repo) => (
-          <List.Item
-            key={repo.fullPath}
-            id={repo.id}
-            title={repo.name}
-            icon={repo.icon}
-            accessoryTitle={tildifyPath(repo.fullPath)}
-            keywords={[repo.name]}
-            actions={<SearchProjectActionPanel repo={repo} preferences={preferences} />}
-          />
-        ))}
-      </List.Section>
-
-      <List.Section title={response?.all?.sectionTitle}>
-        {response?.all?.repos?.map((repo) => (
-          <List.Item
-            key={repo.fullPath}
-            id={repo.id}
-            title={repo.name}
-            icon={repo.icon}
-            accessoryTitle={tildifyPath(repo.fullPath)}
-            keywords={[repo.name]}
-            actions={<SearchProjectActionPanel repo={repo} preferences={preferences} />}
-          />
-        ))}
-      </List.Section>
+      {response?.pinned && getListSection("pinned", response?.pinned)}
+      {response?.recent && getListSection("recent", response?.recent)}
+      {response?.all && getListSection("all", response?.all)}
     </List>
   );
 }